Update node_helper.js
Code efficiency refactoring. Moved ynabBudgetId to be a member variable to the function. And moved ynabAPI to be a member variable to the function so that you don't have to create multiple copies of the ynabAPI in memory. Now there is only the parent single ynabAPI that all of the member methods can access.